Moved commands to their own file
